dios

+++ Cochecito/src/poo4/curso.java
@@ -13,7 +13,7 @@ public class curso {
 	public double getNota() {
 		return nota;
 	}
-	public void setNota(int nota) {
+	public void setNota(double nota) {
 		this.nota = nota;
 	}
 }
